新增:评估report完成
- **人工评估 round4/5/6 聚合结果入档**：`tests/TESTING_STRATEGY.md` 新增 §11.6.4「Phase 3 人工评估 round4/5/6 聚合基线」，记录 2026-06-30 聚合结果（9 条记录、总体 4.61、各维度与 task_type 均值、轮次对比）；`README.md`「质量保障」表同步标注人工评估 round4/5/6 总体均分 4.61（目标 ≥ 3.5）。

+++ app/evaluation/models.py
@@ -109,7 +109,7 @@ class ManualDimensionScore:
     """单维度人工评分"""
 
     dimension: str
-    score: int
+    score: float
     comment: str = ""
 
     def to_dict(self) -> dict[str, Any]:
@@ -150,7 +150,7 @@ class ManualEvaluationRecord:
         scores = [
             ManualDimensionScore(
                 dimension=s["dimension"],
-                score=int(s["score"]),
+                score=float(s["score"]),
                 comment=s.get("comment", ""),
             )
             for s in data["scores"]
